+++ schemes/dc/group_dc/src/PiDCGroup.cpp
@@ -202,7 +202,7 @@ bool PVHSSElg2_ACC_TEST(int msg_num, int degree_f)
     cout << "True result: " << y_native << endl;
     cout << "Eval result: " << y_dig << endl;
     y_eval = conv<ZZ>(y_dig);
-    const ZZ decode_limit = conv<ZZ>(PVHSS_M_MAX);
+    const ZZ decode_limit = conv<ZZ>((unsigned long)PVHSS_M_MAX);
     if (y_native >= decode_limit)
     {
         cout << "P5 accuracy check relaxed: native result is outside DecPed 32-bit decryption range." << endl;
