+++ schemes/dc/rlwe_dc/include/scheme_rlwe_dc.h
@@ -63,7 +63,7 @@ struct SchemeRlweDc {
         dig_t y; pvhss::rlwe::dc::Decode(y,o0.proof,o1.proof,pp.param.f_sk); return NTL::conv<NTL::ZZ>((long)y);
     }
     static bool CanDecodeReference(const SetupOutput&, const NTL::ZZ& reference) {
-        return reference >= 0 && reference < NTL::conv<NTL::ZZ>(PVHSS_M_MAX);
+        return reference >= 0 && reference < NTL::conv<NTL::ZZ>((unsigned long)PVHSS_M_MAX);
     }
 };
 }}
